"Updated log messages, added middleware checks, and modified archive and mail handlers; refactored service and model code."
