+++ data/app.log
@@ -89,3 +89,8 @@ time=2024-11-16T02:16:39.499+05:00 level=INFO msg=Request From=127.0.0.1:62663 M
 time=2024-11-16T02:17:09.036+05:00 level=INFO msg=Request From=127.0.0.1:62669 Method=POST URL=/api/archive/files Additional=""
 time=2024-11-16T02:17:09.036+05:00 level=ERROR msg="this mimetype is not supported. Only: application/vnd.openxmlformats-officedocument.wordprocessingml.document | application/xml | image/jpeg | image/png"
 time=2024-11-16T02:40:45.726+05:00 level=INFO msg=Request From=127.0.0.1:63031 Method=POST URL=/api/mail/file Additional=""
+time=2024-11-16T02:45:46.162+05:00 level=INFO msg=Request From=127.0.0.1:63137 Method=POST URL=/api/mail/file Additional=""
+time=2024-11-16T02:47:04.425+05:00 level=INFO msg=Request From=127.0.0.1:63167 Method=POST URL=/api/mail/file Additional="Middleware check"
+time=2024-11-16T02:47:11.567+05:00 level=INFO msg=Request From=127.0.0.1:63170 Method=POST URL=/api/mail/file Additional="Middleware check"
+time=2024-11-16T02:47:11.568+05:00 level=INFO msg=Request From=127.0.0.1:63170 Method=POST URL=/api/mail/file Additional=""
+time=2024-11-16T02:47:19.233+05:00 level=INFO msg=Request From=127.0.0.1:63172 Method=POST URL=/api/mail/file Additional="Middleware check"
